components/backend/requirements.txt
alembic==1.10.4
amqp==5.1.1
attrs==23.1.0
cffi==1.15.1
croniter==1.3.14
cryptography==40.0.2
exceptiongroup==1.1.1
falcon==3.1.1
flake8==6.0.0
Flake8-pyproject==1.2.3
greenlet==2.0.2
gunicorn==20.1.0
iniconfig==2.0.0
isort==5.12.0
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.7.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-rapidjson==1.10
six==1.16.0
spectree==1.1.2
SQLAlchemy==2.0.12
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
vine==5.0.0
yapf==0.33.0
